refactor(inbox): rewire InboxDetail onto the extracted modules (#994)
Deletes the six inlined definitions now living in inboxDetailHelpers.ts,
useInboxReclassifyV2.ts and FileInspector.tsx, and imports them instead —
the rewiring the parked commit deliberately left undone.

Also lifts all reclassify UI state into useInboxReclassifyState.ts: the
per-file override flow, multi-select + bulk overrides, the #611
heterogeneous-selection acknowledgement gate, and the bulk frame-type undo.
`inspectedIdx` deliberately stays in the component — it is Files-popover
state, not reclassify state.

1472 -> 1026 lines. Behaviour-preserving; 202 inbox tests pass unchanged.
